+++ docs/changes.rst
@@ -25,6 +25,7 @@ v1.1.6 (in development)
 - Update: ``bgobt`` is no longer published.
 - Update: ``boasas`` is no longer published.
 - Update: ``bunny`` is no longer published.
+- Update: ``carpediem`` is no longer published.
 - Update: ``dieselsweetiesweb`` to be more robust to missing elements in the
   feed.